mailhub-recruiter/requirements.txt
agentmail==0.0.60
anthropic==0.68.1
python-dotenv==1.0.0